+++ transaction.cpp
@@ -151,7 +151,7 @@ void Ctransaction::send_error(data_ref r, const std::string& msg)
 
 void Ctransaction::send(data_ref b)
 {
-	if (m_s.sendto(b, reinterpret_cast<const sockaddr*>(&m_a), sizeof(sockaddr_in)) != b.size())
+	if (m_s.sendto(b, reinterpret_cast<const sockaddr*>(&m_a), sizeof(sockaddr_in)) != (ssize_t)b.size())
 		std::cerr << "send failed: " << Csocket::error2a(WSAGetLastError()) << std::endl;
 	srv_stats().sent_udp++;
 }
